pyproject.toml
[tool.poetry]
name = "ybnb2"
version = "0.1.0"
description = "ybnb2"
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.7.3"
nonebot2 = "^2.0.0-beta.1"

[tool.poetry.dev-dependencies]
nb-cli = "^0.6.0"

[tool.nonebot]
plugins = ["nonebot_plugin_directlinker", "nonebot_plugin_repeater", "haruka_bot", "nonebot_plugin_heisi", "nonebot_plugin_petpet"]
plugin_dirs = ["src/plugins"]

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
